Documents for Type Effectiveness Project
Main.java - File for reading in all of the data for each Pokemon for
each text file.
package-info.java - Default file generated by eclipse.
PointlessWindow.java - When opening a window to select a file, this
window needs to be present in order to open (it’s stupid but what can
you do?)
Pokemon.java - A class for Pokemon
PokemonRelease1.txt - A text file used for reading in the data for the
pokemon
Element.java - because Java doesn’t like the word type because it
refers to too many other things in the Java language.
UserInterface.java - blank for right now but for the GUI that we plan
to use